refactor: migrate DataBricksSecret implementation to new integration structure and update tests
